moved parameter functions to parameter.py
